Necessary changes to test csv and sample submission
Changes applied to the test csv data:
I deleted the commodoties column and date column.
I added a column for month and added dummy variables in place of the categorical data.
The Container column has a value that is not in the train set, this caused the test set containing dummies to have an extra column.
I removed this column to fit our model.